chore: - fix mac not support * in build script - move packages/db-client/bridge  to packages/db-client-bridge as one of the monorepo - remove to avoid lint error - add rust folders to .gitignore
