+++ packages/db-client-bridge/lib/index.mjs
@@ -1,0 +1,2 @@
+// This module is the ESM entry point for the library.
+export * from './index.cjs';
