+++ packages/db-client-bridge/lib/load.d.cts
@@ -1,0 +1,1 @@
+export {};
