+++ packages/db-client-bridge/lib/src/index.mjs
@@ -1,0 +1,4 @@
+"use strict";
+// This module is the ESM entry point for the library.
+export * from './index.cjs';
+//# sourceMappingURL=index.mjs.map
