+++ packages/db-client-bridge/lib/src/load.cjs
@@ -1,0 +1,19 @@
+"use strict";
+// This module loads the platform-specific build of the addon on
+// the current system. The supported platforms are registered in
+// the `platforms` object below, whose entries can be managed by
+// by the Neon CLI:
+//
+//   https://www.npmjs.com/package/@neon-rs/cli
+Object.defineProperty(exports, "__esModule", { value: true });
+module.exports = require('@neon-rs/load').proxy({
+    platforms: {
+        'win32-x64-msvc': () => require('@eventstore-rs-js-bridge/win32-x64-msvc'),
+        'darwin-x64': () => require('@eventstore-rs-js-bridge/darwin-x64'),
+        'darwin-arm64': () => require('@eventstore-rs-js-bridge/darwin-arm64'),
+        'linux-x64-gnu': () => require('@eventstore-rs-js-bridge/linux-x64-gnu'),
+        'linux-arm64-gnu': () => require('@eventstore-rs-js-bridge/linux-arm64-gnu')
+    },
+    debug: () => require('../index.node')
+});
+//# sourceMappingURL=load.cjs.map
