--- packages/db-client/bridge/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "eventstore-rs-js-bridge"
-version = "0.1.0"
-description = "A NodeJS binding for the EventStore Rust client"
-authors = ["Yorick Laupa <yo.eight@gmail.com>"]
-license = "MIT"
-edition = "2021"
-exclude = ["index.node"]
-
-[lib]
-crate-type = ["cdylib"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-neon = "1"
-lazy_static = "1"
-
-[dependencies.tokio]
-version = "1"
-features = ["full"]
-
-[dependencies.eventstore]
-version = "3.0.0"
