--- packages/db-client/bridge/package.json
@@ -1,61 +1,0 @@
-{
-  "name": "eventstore-rs-js-bridge",
-  "version": "0.1.0",
-  "description": "A NodeJS binding for the EventStore Rust client",
-  "main": "./lib/index.cjs",
-  "scripts": {
-    "test": "tsc &&cargo test",
-    "cargo-build": "tsc &&cargo build --message-format=json-render-diagnostics > cargo.log",
-    "cross-build": "tsc &&cross build --message-format=json-render-diagnostics > cross.log",
-    "postcargo-build": "neon dist < cargo.log",
-    "postcross-build": "neon dist -m /target < cross.log",
-    "debug": "npm run cargo-build --",
-    "build": "npm run cargo-build -- --release",
-    "cross": "npm run cross-build -- --release",
-    "prepack": "tsc &&neon update",
-    "version": "neon bump --binaries platforms && git add .",
-    "release": "gh workflow run release.yml -f dryrun=false -f version=patch",
-    "dryrun": "gh workflow run publish.yml -f dryrun=true"
-  },
-  "author": "Yorick Laupa",
-  "license": "MIT",
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./lib/index.d.mts",
-        "default": "./lib/index.mjs"
-      },
-      "require": {
-        "types": "./lib/index.d.cts",
-        "default": "./lib/index.cjs"
-      }
-    }
-  },
-  "types": "./lib/index.d.cts",
-  "files": [
-    "lib/**/*.?({c,m}){t,j}s"
-  ],
-  "neon": {
-    "type": "library",
-    "org": "@eventstore-rs-js-bridge",
-    "platforms": "common",
-    "load": "./src/load.cts"
-  },
-  "devDependencies": {
-    "@neon-rs/cli": "^0.1.82",
-    "@tsconfig/node20": "^20.1.4",
-    "@types/node": "^20.11.16",
-    "typescript": "^5.3.3"
-  },
-  "dependencies": {
-    "@neon-rs/load": "^0.1.82"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/EventStore/EventStoreDB-Client-Rust.git"
-  },
-  "bugs": {
-    "url": "https://github.com/EventStore/EventStoreDB-Client-Rust/issues"
-  },
-  "homepage": "https://github.com/EventStore/EventStoreDB-Client-Rust#readme"
-}
