--- packages/db-client/bridge/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@tsconfig/node20/tsconfig.json",
-  "compilerOptions": {
-    "module": "node16",
-    "declaration": true,
-    "outDir": "lib",
-  },
-  "exclude": ["lib"]
-}
