+++ packages/test/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "description": "EventStoreDB Client Tests",
   "scripts": {
-    "build": "npm-run-all build:ts*",
+    "build": "npm-run-all \"build:ts*\"",
     "build:ts": "tsc --build",
     "build:watch": "tsc --build --watch",
     "clean": "run-s clean:*",
@@ -40,8 +40,8 @@
     "@opentelemetry/semantic-conventions": "^1.23.0",
     "@types/jest": "^29.5.12",
     "docker-compose": "^0.24.8",
-    "get-port": "^5.1.1",
     "dotenv": "^16.4.5",
+    "get-port": "^5.1.1",
     "npm-run-all": "^4.1.5",
     "shx": "^0.3.4"
   }
