+++ tsconfig.base.json
@@ -7,8 +7,7 @@
     "sourceMap": true,
     "experimentalDecorators": true,
     "useUnknownInCatchVariables": false,
-    "importsNotUsedAsValues": "error",
     "esModuleInterop": false,
-    "noUnusedLocals": true,
+    "noUnusedLocals": true
   }
 }
